pyproject.toml
[project]
name = "derbynames-zappa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-import-export>=4.3.9",
    "django-s3-sqlite>=0.0.5",
    "django-s3-storage>=0.15.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "huggingface-hub>=0.34.4",
    "markdown>=3.8.2",
    "pillow>=11.3.0",
    "zappa>=0.60.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pre-commit>=4.2.0",
]
